feat(oca-bundle): Add support for fetching bundles with all dependecies
if bundle consist of attributes with type reference all references would
be resolved and provide in a form of oca bundle array as a result.
This commit changes the return of that command from single oca bundle
to vec of bundles, where the last one is always the main bundle from
where it started.